Content refactoring start - checkStatus POC
