+++ src/content/Status.Types.js
@@ -1,0 +1,5 @@
+/* Status Types */
+export const Status = {
+    PAUSED: "paused",
+    PLAYING: "playing"
+};
